// client/src/interfaces/EmployeeFieldErrors.ts
export interface EmployeeFieldErrors {
  first_name?: string[];
  middle_name?: string[];
  last_name?: string[];
  suffix_name?: string[];
  birth_date?: string[];
  gender?: string[];
  address?: string[];
  contact_number?: string[];
  role?: string[];
  email?: string[];
  password?: string[];
  password_confirmation?: string[];
}